add sync and whisper loops

+++ src/cli_opts.rs
@@ -3,7 +3,7 @@ use clap::Clap;
 #[derive(Clap)]
 pub struct CliOpts {
     #[clap(long)]
-    pub period: i32,
+    pub period: u64,
     #[clap(long)]
     pub port: String,
     #[clap(long)]
